Transition to `State::Closed` on error.
Client code may continue to invoke `Sink` methods after encountering an
error. The current implementation does not transition to a different
state on errors which may cause the repeated invocations to poll the
future after completion.

To mitigate this, we now transition to `State::Closed` immediately if an
error occurs. This means that clients may then call `poll_flush` or
`poll_close` again, but without effect. If on the other hand, client
code attempts to keep sending items to the sink via `poll_ready` or
`start_send` we enter panic mode as the sink can not be used once it
experienced an error.