RFC: Add `Error` type.
Originally, the `SinkImpl` was intended to be generic over the error
type of the inner future in order to not require any kind of error
mapping or `From` impls. Instead, whenever the `SinkImpl` is used in
a "wrong way", a panic happens. This however makes it more difficult
to use `SinkImpl` in certain contexts, e.g. concurrently using locks,
when for instance one thread successfully closes the sink using
`poll_close` and another one calls `poll_ready` afterwards.

Therefore, this PR is an RFC to add an error type to indicate to callers
when the sink is closed so they can handle it more gracefully. It will
require those error type mappings but usage should become more robust.