Remove `Error`.
In https://github.com/paritytech/quicksink/pull/4 an error type has
been added instead of `panic!` which this PR reverts. While the
original motivation is still valid it did not work out in practice.
For instance the fact that `SinkExt::sink_map_err` works with
`FnOnce` only makes repeated error mapping difficult which defeats
the purpose of `Error`.

Therefore this PR goes back to being generic over errors and produces
panics when the sink is used wrongly, i.e. if its methods are called
again after an error occured or if after `poll_close` other methods
are called. Compared to 0.1.1 the only difference in behaviour is that
when closing we already panic if `poll_ready` is called and not just
when `start_send` is called since an ok from `poll_ready` would
indicate that `start_send` could be called which is really not the
case.